+++ src/paco/cookiecutters/privatepypi/__init__.py
@@ -3,7 +3,7 @@ from paco.models import vocabulary
 
 def get_project_context(paco_ctx):
     return {
-        'project_name': os.path.basename(os.path.normpath(paco_ctx.home)),
+        'project_name': os.path.basename(str(paco_ctx.home)),
         'budget': 'Y',
         'budget_allowed_values': ['Y','y','yes','N','n','no'],
         'project_title': 'My Paco Project',