+++ src/paco/cookiecutters/s3lambda/__init__.py
@@ -3,7 +3,7 @@ from paco.models import vocabulary
 
 def get_project_context(paco_ctx):
     return {
-        'project_name': os.path.basename(os.path.normpath(paco_ctx.home)),
+        'project_name': os.path.basename(str(paco_ctx.home)),
         'project_title': None,
         'admin_email': None,
         'network_environment_name': None,